Restructured project folder for better organization
