+++ src/components/common/Notifications.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { socket } from "../socket";
+import { socket } from "../../socket";
 
 const Notifications = () => {
   const [notifications, setNotifications] = useState([]);
