--- src/pages/ContactUs.css
@@ -1,84 +1,0 @@
-/* ContactUs.css */
-
-.contact-page {
-    font-family: Arial, sans-serif;
-    text-align: center;
-    color: #333;
-    background-color: #f9f9f9;
-    padding: 20px;
-  }
-  
-  .contact-header {
-    background-color: #6c63ff;
-    color: white;
-    padding: 50px 20px;
-    border-radius: 5px;
-  }
-  
-  .contact-form {
-    background: white;
-    padding: 30px;
-    margin: 20px auto;
-    max-width: 600px;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  }
-  
-  .form-group {
-    display: flex;
-    gap: 20px;
-    margin-bottom: 15px;
-  }
-  
-  .form-group input, textarea {
-    width: 100%;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-  }
-  
-  button {
-    background: #6c63ff;
-    color: white;
-    border: none;
-    padding: 10px 20px;
-    font-size: 1rem;
-    cursor: pointer;
-    border-radius: 5px;
-  }
-  
-  button:hover {
-    background: #534dbd;
-  }
-  
-  .map {
-    margin: 30px auto;
-    max-width: 800px;
-  }
-  
-  .map img {
-    width: 100%;
-    border-radius: 10px;
-  }
-  
-  .faq {
-    margin-top: 50px;
-    padding: 20px;
-  }
-  
-  .faq-item {
-    background: white;
-    padding: 15px;
-    border-radius: 5px;
-    margin: 10px auto;
-    max-width: 500px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  }
-  
-  .footer {
-    background: #333;
-    color: white;
-    padding: 20px;
-    margin-top: 50px;
-  }
-
