--- src/pages/Footer.css
@@ -1,67 +1,0 @@
-.footer {
-    background-color: #1a121a;
-    color: white;
-    padding: 40px;
-  }
-  
-  .footer-subscribe {
-    text-align: center;
-    margin-bottom: 30px;
-  }
-  
-  .footer-subscribe p {
-    font-size: 18px;
-    margin-bottom: 10px;
-  }
-  
-  .subscribe-box {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: white;
-    border-radius: 5px;
-    overflow: hidden;
-    max-width: 400px;
-    margin: 0 auto;
-  }
-  
-  .subscribe-box input {
-    border: none;
-    padding: 10px;
-    flex: 1;
-    outline: none;
-  }
-  
-  .subscribe-box button {
-    background-color: #4c2a4a;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    cursor: pointer;
-  }
-  
-  .footer-links {
-    display: flex;
-    justify-content: space-between;
-    flex-wrap: wrap;
-    margin-top: 20px;
-  }
-  
-  .footer-links h4 {
-    margin-bottom: 10px;
-  }
-  
-  .footer-links ul {
-    list-style: none;
-    padding: 0;
-  }
-  
-  .footer-links ul li {
-    margin-bottom: 5px;
-  }
-  
-  .footer-links ul li a {
-    color: white;
-    text-decoration: none;
-  }
-
