--- src/pages/Navbar.css
@@ -1,62 +1,0 @@
-.navbar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 15px 40px;
-    background-color: white;
-    border-bottom: 1px solid #ddd;
-  }
-  
-  .navbar-left {
-    display: flex;
-    align-items: center;
-  }
-  
-  .logo {
-    width: 40px;
-    height: 40px;
-    margin-right: 10px;
-  }
-  
-  .logo-text {
-    font-size: 18px;
-    font-weight: bold;
-    color: #4c2a4a;
-    font-style: italic;
-  }
-  
-  .nav-links {
-    display: flex;
-    list-style: none;
-    gap: 20px;
-  }
-  
-  .nav-links li a {
-    text-decoration: none;
-    color: #333;
-    font-size: 16px;
-  }
-  
-  .navbar-right {
-    display: flex;
-    gap: 10px;
-  }
-  
-  .join-btn {
-    background-color: #4c2a4a;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 20px;
-    cursor: pointer;
-  }
-  
-  .start-btn {
-    background-color: transparent;
-    color: #4c2a4a;
-    border: 2px solid #4c2a4a;
-    padding: 8px 18px;
-    border-radius: 20px;
-    cursor: pointer;
-  }
-
